+++ src/main/scala/simulator/trader/TraderParams.scala
@@ -1,3 +1,3 @@
 package simulator.trader
 
-case class TraderParams(id: Int, initialBalance: Int, initialHoldings: Int)
+case class TraderParams(id: Int, initialBalance: Double, initialHoldings: Double)
